+++ lib/service_actor/configurable.rb
@@ -15,6 +15,26 @@ module ServiceActor::Configurable
       child.failure_class = failure_class || ServiceActor::Failure
     end
 
-    attr_accessor :argument_error_class, :failure_class
+    def argument_error_class=(value)
+      validate_provided_error_class(value)
+
+      @argument_error_class = value
+    end
+
+    def failure_class=(value)
+      validate_provided_error_class(value)
+
+      @failure_class = value
+    end
+
+    attr_reader :argument_error_class, :failure_class
+
+    private
+
+    def validate_provided_error_class(value)
+      return if value.is_a?(Class) && value <= Exception
+
+      raise ArgumentError, "Expected #{value} to be a subclass of Exception"
+    end
   end
 end