+++ spec/examples/with_unset_output.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class WithUnsetOutput < Actor
+  output :value, type: String, allow_nil: true
+end