+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Fixes:
   are subclasses of `Exception` (#132)
 - Skip `default` check for actor outputs (#135)
 - Ensure provided `fail_on` arguments are subclasses of `Exception` (#136)
+- Ensure `input` and `output` names do not collide with
+  `ServiceActor::Result` instance methods (#138)
 
 ## v3.7.0